111121ghj

+++ handlers/users/downloder.py
@@ -6,10 +6,10 @@ import yt_dlp
 
 
 @dp.message_handler()
-async def download(message: Message):
+async def download(message: types.Message):
     url = message.text
 
-    await message.reply("⏳")
+    await message.reply("⏳ Yuklanmoqda...")
 
     dir_name = "downloads"
     os.makedirs(dir_name, exist_ok=True)
@@ -18,18 +18,22 @@ async def download(message: Message):
         "format": "best",
         "outtmpl": f"{dir_name}/%(title)s.%(ext)s",
         "concurrent_fragment_downloads": 5,
-        "cookies": "/home/ubuntu/Downloader/cookies.txt"
+        "cookiefile": "/home/ubuntu/Downloader/cookies.txt"  # 'cookies' fayli yo'li
     }
 
     try:
         with yt_dlp.YoutubeDL(ydl_opts) as ydl:
-            file = ydl.extract_info(url, download=True)
-            file_name = ydl.prepare_filename(file)
-        with open(file_name, 'rb') as video:
-            await message.answer_video(video)
+            file_info = ydl.extract_info(url, download=True)
+            file_name = ydl.prepare_filename(file_info)
 
-        os.remove(file_name)
+        if os.path.exists(file_name):
+            with open(file_name, 'rb') as video:
+                await message.answer_video(video=types.InputFile(video))
+            os.remove(file_name)
+        else:
+            await message.reply("❌ Yuklab olishda xatolik yuz berdi: Fayl topilmadi.")
 
+    except yt_dlp.utils.DownloadError as e:
+        await message.reply(f"❌ Yuklab olishda xatolik yuz berdi: {str(e)}")
     except Exception as e:
-        logging.exception(e)
-        await message.reply("❌ Xatolik yuz berdi")
+        await message.reply(f"❌ Kutilmagan xatolik yuz berdi: {str(e)}")
